moved all tests to a single file
